refactor(ui): TMS-2209 add index files
Added index files for models, services, and stores so that the imports are easier to remember.
