+++ frontend/src/models/index.ts
@@ -1,0 +1,3 @@
+export { Role } from './role.model'
+export { Tenant } from './tenant.model'
+export { User } from './user.model'
