+++ frontend/src/services/index.ts
@@ -1,0 +1,3 @@
+export { roleService } from './role.service'
+export { tenantService } from './tenant.service'
+export { userService } from './user.service'
