+++ frontend/src/stores/useTenantStore.ts
@@ -1,10 +1,8 @@
 import { defineStore } from 'pinia'
 import { ref } from 'vue'
 
-import { Role } from '@/models/role.model'
-import { Tenant } from '@/models/tenant.model'
-import { User } from '@/models/user.model'
-import { tenantService } from '@/services/tenant.service'
+import { Role, Tenant, User } from '@/models'
+import { tenantService } from '@/services'
 
 export const useTenantStore = defineStore('tenant', () => {
   const tenants = ref<Tenant[]>([])
